% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Set up a Database}
\usage{
db_connect(name, path = ".sqlite_data", engine = RSQLite::SQLite())
}
\arguments{
\item{name}{The name of a data set}

\item{path}{The path where to store file database, if applicable.}

\item{engine}{A DBI-compatible database engine.}
}
\value{
con A \link[DBI:DBIConnection-class]{DBIConnection}.
}
\description{
Set up a Database
}
