% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{import_validpairs}
\alias{import_validpairs}
\title{Import a Valid Pair File to Database}
\usage{
import_validpairs(con, file, debug = TRUE)
}
\arguments{
\item{con}{A DBIConnection.}

\item{file}{An tab-delimited assignment file readable
using \code{\link[=read_validpairs]{read_validpairs()}}.}

\item{debug}{Should debug output be produced or not?}
}
\value{
(invisibly) A \link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi} object.
}
\description{
Import a Valid Pair File to Database
}
