% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md5_file.R
\name{md5_file}
\alias{md5_file}
\title{Get a file MD5 checksum and output to paired *.md5 file}
\usage{
md5_file(pathname)
}
\arguments{
\item{pathname}{File for which MD5 checksum should be produced}
}
\value{
A MD5 checksum string validated to consist of 32 hexadecimal
characters (if not, an error is produced).
}
\description{
Get a file MD5 checksum and output to paired *.md5 file
}
\details{
If a paired \file{*.md5} checksum file already exists and is newer than
the source \code{pathname} file, then the checksum is read from the checksum
file.  In all other cases, a new MD5 checksum is calculated and a paired
\file{*.md5} file is written/updated.
}
