% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{nrow2}
\alias{nrow2}
\title{Get the number of rows in a database table}
\usage{
nrow2(tbl)
}
\arguments{
\item{tbl}{A \link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi} object.}
}
\value{
An integer.
}
\description{
Get the number of rows in a database table
}
\references{
Implementation of \code{nrow2()} is from
\url{http://www.win-vector.com/blog/2017/09/it-is-needlessly-difficult-to-count-rows-using-dplyr/}
}
