% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_assignments.R
\name{read_assignments}
\alias{read_assignments}
\title{Read Assignments Files}
\usage{
read_assignments(file, columns = NULL, ...)
}
\arguments{
\item{file}{Pathname to a \file{*_assignments.txt(.gz)} file.}

\item{columns}{(optional) Name of columns to be read.}

\item{...}{Additional arguments passed to \code{\link[readr:read_tsv]{readr::read_tsv()}}.}
}
\value{
A data.frame with four columns:
\item{\code{readname}}{...}
\item{\code{left_inner_barcode}}{...}
\item{\code{right_inner_barcode}}{...}
\item{\code{outer_barcode}}{...}
}
\description{
An 'assignments' file ...
}
\details{
The description here are adopted from \href{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE84nnn/GSE84920/suppl/GSE84920\%5FREADME\%2Etxt}{GSE84920_README.txt}.
It is sparse on what 'assignments' files are, but it says:
"TXT [ed. we think they meant ASSIGNMENTS] files represent the barcode
associations for each sequenced read where two barcodes were identifiable.
The format is \code{READNAME\\tLEFT_INNER_BARCODE\\tRIGHT_INNER_BARCODE\\tOUTER_BARCODE},
where \code{LEFT_INNER_BARCODE} and \code{RIGHT_INNER_BARCODE} should match for
valid reads.  We include all reads here, but note that for all analyses
presented in the associated paper, all reads where these barcodes do not
match were discarded."

[ed. we think they meant ASSIGNMENTS]: R:ed.%20we%20think%20they%20meant%20ASSIGNMENTS%5C
}
\section{Validation}{

The \code{read_assignments()} function does some basic validation on the
values read.
}

\examples{
path <- system.file("extdata", package = "ramani")
file <- file.path(path, "GSM2254215_ML1.rows=1-1000_assignments.txt.gz")
data <- read_assignments(file)
print(data)

}
