% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genomesize.R
\name{read_genomesize}
\alias{read_genomesize}
\title{Read Genome-Size Files}
\usage{
read_genomesize(file, ...)
}
\arguments{
\item{file}{Pathname to a \file{*.genomesize} file.}

\item{...}{Additional arguments passed to \code{\link[readr:read_tsv]{readr::read_tsv()}}.}
}
\value{
A data.frame with four columns:
\item{\code{chr}}{Chromosome label}
\item{\code{length}}{Number of nucleotides in chromosome}
}
\description{
Read Genome-Size Files
}
\details{
This file originates from \url{https://github.com/VRam142/combinatorialHiC}.
}
\section{Validation}{

The \code{read_genomsize()} function does some basic validation on the
values read.
}

\examples{
path <- system.file("extdata", package = "ramani")
file <- file.path(path, "combo_hg19_mm10.genomesize")
data <- read_genomesize(file)
print(data)

}
