% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_percentages.R
\name{read_percentages}
\alias{read_percentages}
\title{Read Percentages Files}
\usage{
read_percentages(file, col_names = FALSE, ...)
}
\arguments{
\item{file}{Pathname to a *.percentages(.tar.gz) file}
}
\value{
A data.frame
}
\description{
Read Percentages Files
}
\details{
The `read_percentages()` function does some basic validation on the
values read.
}
