% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{import_assignments}
\alias{import_assignments}
\title{Import an Assignments File to Database}
\usage{
import_assignments(con, file, debug = TRUE)
}
\arguments{
\item{con}{A DBIConnection.}

\item{file}{An tab-delimited assignment file readable
using \code{\link[=read_assignments]{read_assignments()}}.}

\item{debug}{Should debug output be produced or not?}
}
\value{
(invisibly) A \link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi} object.
}
\description{
Import an Assignments File to Database
}
\examples{
name <- "GSM2254215_ML1.rows=1-1000"
con <- db_connect(name, path = file.path(tempdir(), ".sqlite_data"))

path <- system.file("extdata", package = "ramani")
import_assignments(con, file = file.path(path, sprintf("\%s_assignments.txt.gz", name)))

data <- dplyr::tbl(con, "assignments")
print(data)

import_validpairs(con, file = file.path(path, sprintf("\%s.validPairs.txt.gz", name)))

data <- dplyr::tbl(con, "validpairs")
print(data)

file <- file.path(path, sprintf("\%s.percentages.txt.gz", name))
print(file.size(file))  ## 273357
data <- read_percentages(file)  ## Takes 1-2 sec to read (270 kB)
print(data)

DBI::dbDisconnect(con)
}
