% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genomesize.R
\name{read_genomesize}
\alias{read_genomesize}
\title{Read Genome-Size Files}
\usage{
read_genomesize(file, ...)
}
\arguments{
\item{file}{Pathname to a \file{*.genomesize} file.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_tsv()}}.}
}
\value{
A data.frame with four columns:
\item{\code{chr}}{Chromosome label}
\item{\code{length}}{Number of nucleotides in chromosome}
}
\description{
Read Genome-Size Files
}
\details{
This file originates from \url{https://github.com/VRam142/combinatorialHiC}.
}
\section{Validation}{

The \code{read_genomsize()} function does some basic validation on the
values read.
}

\examples{
path <- system.file("extdata", package = "GSE84920.parser")
file <- file.path(path, "combo_hg19_mm10.genomesize")
data <- read_genomesize(file)
print(data)
# # A tibble: 43 x 2
#    chr            length
#    <chr>           <int>
#  1 human_chr1  249250621
#  2 human_chr2  243199373
#  3 human_chr3  198022430
#  4 human_chr4  191154276
#  5 human_chr5  180915260
#  6 human_chr6  171115067
#  7 human_chr7  159138663
#  8 human_chr8  146364022
#  9 human_chr9  141213431
# 10 human_chr10 135534747
# # … with 33 more rows

}
