% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_percentages.R
\name{read_percentages}
\alias{read_percentages}
\title{Read Percentages Files}
\usage{
read_percentages(file, columns = NULL, ...)
}
\arguments{
\item{file}{Pathname to a \file{*.percentages.txt(.gz)} file.}

\item{columns}{(optional) Name of columns to be read.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{readr::read_tsv()}}.}
}
\value{
A data.frame with 17 columns:
\item{\code{hg19_frac}}{Fraction Reads Mapping to hg19}
\item{\code{mm10_frac}}{Fraction Reads Mapping to mm10}
\item{\code{hg19_count}}{Number of Reads Mapping to hg19}
\item{\code{mm10_count}}{Number of Reads Mapping to mm10}
\item{\code{hg19mm10_count}}{Total Number of Read Pairs filtering out
interspecies (== \code{read_hg19_count} + \code{read_mm10_count})}
\item{\code{pair_count}}{Total Number of Reads Pairs}
\item{\code{inner_barcode}}{Round 1 Barcode (Inner)}
\item{\code{outer_barcode}}{Round 2 Barcode (Outer)}
\item{\code{is_observed}}{\code{True} or \code{Randomized};
\code{True} are the observed data,
\code{Randomized} are the result of shuffling all
barcode assignments to new reads}
\item{\code{Col10}}{\code{All} or \code{Long};
\code{All} are all intrachromosomal / interchromosomal reads
associated with a cellular index;
\code{Long} are only inter- and intra > 20 kbp}
\item{\code{dpnii_1x}}{Number of times a DpnII fragment is observed once}
\item{\code{dpnii_2x}}{Number of times a DpnII fragment is observed twice}
\item{\code{dpnii_3x}}{Number of times a DpnII fragment is observed thrice}
\item{\code{dpnii_4x}}{Number of times a DpnII fragment is observed four times}
\item{\code{cistrans_ratio}}{Cis-trans ratio for that cellular index}
\item{\code{hela_allele_frac}}{If applicable (only HeLa S3 and HAP1 cells),
fraction of homozygous alternate HeLa allele calls}
\item{\code{celltype}}{Programmed cell type assignment, if applicable.
For ML libraries, only cells with >0.95 in \code{hg19_frac} or
\code{mm10_frac} receive assignments.}
}
\description{
A 'percentages' file provides summary statistics for each processed cellular
index in combinatorial single-cell Hi-C data.
}
\details{
The description here are adopted from \href{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE84nnn/GSE84920/suppl/GSE84920\%5FREADME\%2Etxt}{GSE84920_README.txt}, which
refers to the "manuscript" for description of the "PERCENTAGES" files.
The column names returned are ours, because the data files do not provide
column names.
}
\section{Validation}{

The \code{read_percentages()} function does some basic validation on the
values read.
}

\examples{
path <- system.file("extdata", package = "GSE84920.parser")
file <- file.path(path, "GSM2254215_ML1.rows=1-1000.percentages.txt.gz")
data <- read_percentages(file)
print(data)
# # A tibble: 1,000 x 16
#    hg19_frac mm10_frac hg19_count mm10_count pair_count inner_barcode outer_barcode is_observed
#        <dbl>     <dbl>      <int>      <int>      <int> <chr>         <chr>         <chr>      
#  1     1.000 0.0000356      28050          1      28052 ACCACCAC      TCAGATGC      True       
#  2     0.680 0.320          19081       8970      28052 ACCACCAC      TCAGATGC      Randomized 
#  3     1.000 0.0000370      27010          1      28052 ACCACCAC      TCAGATGC      True       
#  4     0.683 0.317          18444       8555      28052 ACCACCAC      TCAGATGC      Randomized 
#  5     0     1                  0          1          1 CATAGCGC      ACTTGATA      True       
#  6     1     0                  1          0          1 CATAGCGC      ACTTGATA      Randomized 
#  7     0     1                  0          1          1 CATAGCGC      ACTTGATA      True       
#  8     1     0                  1          0          1 CATAGCGC      ACTTGATA      Randomized 
#  9     1     0                  2          0          2 GGCCGTTC      GCCATTAA      True       
# 10     1     0                  2          0          2 GGCCGTTC      GCCATTAA      Randomized 
# # … with 990 more rows, and 8 more variables: Col10 <chr>, dpnii_1x <int>, dpnii_2x <int>,
# #   dpnii_3x <int>, dpnii_4x <int>, cistrans_ratio <dbl>, hela_allele_frac <dbl>,
# #   celltype <chr>

print(table(data$celltype))
###   HAP1    HeLa    MEF   Patski Undetermined 
###    156     163    174      152          355 


}
