% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validpairs.R
\name{read_validpairs}
\alias{read_validpairs}
\title{Read Valid-Pairs Files}
\usage{
read_validpairs(file, ...)
}
\arguments{
\item{file}{Pathname to a \file{*.validPairs.txt(.gz)} file.}

\item{...}{Additional arguments passed to \code{\link[readr:read_tsv]{readr::read_tsv()}}.}
}
\value{
A data.frame with 17 columns:
\item{\code{???}}{???}
}
\description{
A 'VALIDPAIRS' file is for all valid pairs in each library.
}
\details{
The description here are adopted from \href{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE84nnn/GSE84920/suppl/GSE84920\%5FREADME\%2Etxt}{GSE84920_README.txt}.
It is sparse on what 'VALIDPAIRS' files contain, but says that they
"VALIDPAIRS files are for all valid pairs in each library.
BEDS were paired based on identical read name and simultaneously
associated with the proper cellular index. VALIDPAIRS also include the
result of bedtools closest, to determine the closest DpnII restriction
site and distance for each mate."
}
\section{Validation}{

The \code{read_validpairs()} function does some basic validation on the
values read.
}

\examples{
path <- system.file("extdata", package = "ramani")
file <- file.path(path, "GSM2254215_ML1.rows=1-1000.validPairs.txt.gz")
data <- read_validpairs(file)
print(data)

}
